Updated naming conventions
